Avoid creating circular dependencies in drive folders
